refactor: move visualizer page creation to a separate function and rename exports
