+++ src/ReactVisualizer.jsx
@@ -1,0 +1,5 @@
+'use strict';
+
+const React = require('react');
+
+
