+++ src/visualizer.js
@@ -20,7 +20,7 @@ const jquery = "<script>\n        /*! jQuery v2.1.4 | (c) 2005, 2015 jQuery Foun
 const urijs = "\n    \n    <script>\n        /*! URI.js v1.16.1 http://medialize.github.io/URI.js/ */\n        /* build contains: URI.js, URI.fragmentQuery.js */\n        (function(f,g){\"object\"===typeof exports?module.exports=g(require(\"./punycode\"),require(\"./IPv6\"),require(\"./SecondLevelDomains\")):\"function\"===typeof define&&define.amd?define([\"./punycode\",\"./IPv6\",\"./SecondLevelDomains\"],g):f.URI=g(f.punycode,f.IPv6,f.SecondLevelDomains,f)})(this,function(f,g,q,n){function d(a,b){var c=1<=arguments.length,h=2<=arguments.length;if(!(this instanceof d))return c?h?new d(a,b):new d(a):new d;if(void 0===a){if(c)throw new TypeError(\"undefined is not a valid argument for URI\");\n            a=\"undefined\"!==typeof location?location.href+\"\":\"\"}this.href(a);return void 0!==b?this.absoluteTo(b):this}function l(a){return a.replace(/([.*+?^=!:${}()|[\\]\\/\\\\])/g,\"\\\\$1\")}function m(a){return void 0===a?\"Undefined\":String(Object.prototype.toString.call(a)).slice(8,-1)}function k(a){return\"Array\"===m(a)}function z(a,b){var c={},d,p;if(\"RegExp\"===m(b))c=null;else if(k(b))for(d=0,p=b.length;d<p;d++)c[b[d]]=!0;else c[b]=!0;d=0;for(p=a.length;d<p;d++)if(c&&void 0!==c[a[d]]||!c&&b.test(a[d]))a.splice(d,\n                1),p--,d--;return a}function w(a,b){var c,d;if(k(b)){c=0;for(d=b.length;c<d;c++)if(!w(a,b[c]))return!1;return!0}var p=m(b);c=0;for(d=a.length;c<d;c++)if(\"RegExp\"===p){if(\"string\"===typeof a[c]&&a[c].match(b))return!0}else if(a[c]===b)return!0;return!1}function A(a,b){if(!k(a)||!k(b)||a.length!==b.length)return!1;a.sort();b.sort();for(var c=0,d=a.length;c<d;c++)if(a[c]!==b[c])return!1;return!0}function C(a){return escape(a)}function x(a){return encodeURIComponent(a).replace(/[!'()*]/g,C).replace(/\\*/g,\n                \"%2A\")}function u(a){return function(b,c){if(void 0===b)return this._parts[a]||\"\";this._parts[a]=b||null;this.build(!c);return this}}function B(a,b){return function(c,d){if(void 0===c)return this._parts[a]||\"\";null!==c&&(c+=\"\",c.charAt(0)===b&&(c=c.substring(1)));this._parts[a]=c;this.build(!d);return this}}var D=n&&n.URI;d.version=\"1.16.1\";var e=d.prototype,r=Object.prototype.hasOwnProperty;d._parts=function(){return{protocol:null,username:null,password:null,hostname:null,urn:null,port:null,path:null,\n            query:null,fragment:null,duplicateQueryParameters:d.duplicateQueryParameters,escapeQuerySpace:d.escapeQuerySpace}};d.duplicateQueryParameters=!1;d.escapeQuerySpace=!0;d.protocol_expression=/^[a-z][a-z0-9.+-]*$/i;d.idn_expression=/[^a-z0-9\\.-]/i;d.punycode_expression=/(xn--)/i;d.ip4_expression=/^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$/;d.ip6_expression=/^\\s*((([0-9A-Fa-f]{1,4}:){7}([0-9A-Fa-f]{1,4}|:))|(([0-9A-Fa-f]{1,4}:){6}(:[0-9A-Fa-f]{1,4}|((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3})|:))|(([0-9A-Fa-f]{1,4}:){5}(((:[0-9A-Fa-f]{1,4}){1,2})|:((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3})|:))|(([0-9A-Fa-f]{1,4}:){4}(((:[0-9A-Fa-f]{1,4}){1,3})|((:[0-9A-Fa-f]{1,4})?:((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){3}(((:[0-9A-Fa-f]{1,4}){1,4})|((:[0-9A-Fa-f]{1,4}){0,2}:((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){2}(((:[0-9A-Fa-f]{1,4}){1,5})|((:[0-9A-Fa-f]{1,4}){0,3}:((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){1}(((:[0-9A-Fa-f]{1,4}){1,6})|((:[0-9A-Fa-f]{1,4}){0,4}:((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3}))|:))|(:(((:[0-9A-Fa-f]{1,4}){1,7})|((:[0-9A-Fa-f]{1,4}){0,5}:((25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)(\\.(25[0-5]|2[0-4]\\d|1\\d\\d|[1-9]?\\d)){3}))|:)))(%.+)?\\s*$/;\n            d.find_uri_expression=/\\b((?:[a-z][\\w-]+:(?:\\/{1,3}|[a-z0-9%])|www\\d{0,3}[.]|[a-z0-9.\\-]+[.][a-z]{2,4}\\/)(?:[^\\s()<>]+|\\(([^\\s()<>]+|(\\([^\\s()<>]+\\)))*\\))+(?:\\(([^\\s()<>]+|(\\([^\\s()<>]+\\)))*\\)|[^\\s`!()\\[\\]{};:'\".,<>?\\u00ab\\u00bb\\u201c\\u201d\\u2018\\u2019]))/ig;d.findUri={start:/\\b(?:([a-z][a-z0-9.+-]*:\\/\\/)|www\\.)/gi,end:/[\\s\\r\\n]|$/,trim:/[`!()\\[\\]{};:'\".,<>?\\u00ab\\u00bb\\u201c\\u201d\\u201e\\u2018\\u2019]+$/};d.defaultPorts={http:\"80\",https:\"443\",ftp:\"21\",gopher:\"70\",ws:\"80\",wss:\"443\"};d.invalid_hostname_characters=\n                    /[^a-zA-Z0-9\\.-]/;d.domAttributes={a:\"href\",blockquote:\"cite\",link:\"href\",base:\"href\",script:\"src\",form:\"action\",img:\"src\",area:\"href\",iframe:\"src\",embed:\"src\",source:\"src\",track:\"src\",input:\"src\",audio:\"src\",video:\"src\"};d.getDomAttribute=function(a){if(a&&a.nodeName){var b=a.nodeName.toLowerCase();return\"input\"===b&&\"image\"!==a.type?void 0:d.domAttributes[b]}};d.encode=x;d.decode=decodeURIComponent;d.iso8859=function(){d.encode=escape;d.decode=unescape};d.unicode=function(){d.encode=x;d.decode=\n                    decodeURIComponent};d.characters={pathname:{encode:{expression:/%(24|26|2B|2C|3B|3D|3A|40)/ig,map:{\"%24\":\"$\",\"%26\":\"&\",\"%2B\":\"+\",\"%2C\":\",\",\"%3B\":\";\",\"%3D\":\"=\",\"%3A\":\":\",\"%40\":\"@\"}},decode:{expression:/[\\/\\?#]/g,map:{\"/\":\"%2F\",\"?\":\"%3F\",\"#\":\"%23\"}}},reserved:{encode:{expression:/%(21|23|24|26|27|28|29|2A|2B|2C|2F|3A|3B|3D|3F|40|5B|5D)/ig,map:{\"%3A\":\":\",\"%2F\":\"/\",\"%3F\":\"?\",\"%23\":\"#\",\"%5B\":\"[\",\"%5D\":\"]\",\"%40\":\"@\",\"%21\":\"!\",\"%24\":\"$\",\"%26\":\"&\",\"%27\":\"'\",\"%28\":\"(\",\"%29\":\")\",\"%2A\":\"*\",\"%2B\":\"+\",\"%2C\":\",\",\n                \"%3B\":\";\",\"%3D\":\"=\"}}},urnpath:{encode:{expression:/%(21|24|27|28|29|2A|2B|2C|3B|3D|40)/ig,map:{\"%21\":\"!\",\"%24\":\"$\",\"%27\":\"'\",\"%28\":\"(\",\"%29\":\")\",\"%2A\":\"*\",\"%2B\":\"+\",\"%2C\":\",\",\"%3B\":\";\",\"%3D\":\"=\",\"%40\":\"@\"}},decode:{expression:/[\\/\\?#:]/g,map:{\"/\":\"%2F\",\"?\":\"%3F\",\"#\":\"%23\",\":\":\"%3A\"}}}};d.encodeQuery=function(a,b){var c=d.encode(a+\"\");void 0===b&&(b=d.escapeQuerySpace);return b?c.replace(/%20/g,\"+\"):c};d.decodeQuery=function(a,b){a+=\"\";void 0===b&&(b=d.escapeQuerySpace);try{return d.decode(b?a.replace(/\\+/g,\n                    \"%20\"):a)}catch(c){return a}};var t={encode:\"encode\",decode:\"decode\"},v,y=function(a,b){return function(c){try{return d[b](c+\"\").replace(d.characters[a][b].expression,function(c){return d.characters[a][b].map[c]})}catch(h){return c}}};for(v in t)d[v+\"PathSegment\"]=y(\"pathname\",t[v]),d[v+\"UrnPathSegment\"]=y(\"urnpath\",t[v]);t=function(a,b,c){return function(h){var p;p=c?function(a){return d[b](d[c](a))}:d[b];h=(h+\"\").split(a);for(var e=0,f=h.length;e<f;e++)h[e]=p(h[e]);return h.join(a)}};d.decodePath=\n                    t(\"/\",\"decodePathSegment\");d.decodeUrnPath=t(\":\",\"decodeUrnPathSegment\");d.recodePath=t(\"/\",\"encodePathSegment\",\"decode\");d.recodeUrnPath=t(\":\",\"encodeUrnPathSegment\",\"decode\");d.encodeReserved=y(\"reserved\",\"encode\");d.parse=function(a,b){var c;b||(b={});c=a.indexOf(\"#\");-1<c&&(b.fragment=a.substring(c+1)||null,a=a.substring(0,c));c=a.indexOf(\"?\");-1<c&&(b.query=a.substring(c+1)||null,a=a.substring(0,c));\"//\"===a.substring(0,2)?(b.protocol=null,a=a.substring(2),a=d.parseAuthority(a,b)):(c=a.indexOf(\":\"),\n            -1<c&&(b.protocol=a.substring(0,c)||null,b.protocol&&!b.protocol.match(d.protocol_expression)?b.protocol=void 0:\"//\"===a.substring(c+1,c+3)?(a=a.substring(c+3),a=d.parseAuthority(a,b)):(a=a.substring(c+1),b.urn=!0)));b.path=a;return b};d.parseHost=function(a,b){a=a.replace(/\\\\/g,\"/\");var c=a.indexOf(\"/\"),d;-1===c&&(c=a.length);if(\"[\"===a.charAt(0))d=a.indexOf(\"]\"),b.hostname=a.substring(1,d)||null,b.port=a.substring(d+2,c)||null,\"/\"===b.port&&(b.port=null);else{var p=a.indexOf(\":\");d=a.indexOf(\"/\");\n                p=a.indexOf(\":\",p+1);-1!==p&&(-1===d||p<d)?(b.hostname=a.substring(0,c)||null,b.port=null):(d=a.substring(0,c).split(\":\"),b.hostname=d[0]||null,b.port=d[1]||null)}b.hostname&&\"/\"!==a.substring(c).charAt(0)&&(c++,a=\"/\"+a);return a.substring(c)||\"/\"};d.parseAuthority=function(a,b){a=d.parseUserinfo(a,b);return d.parseHost(a,b)};d.parseUserinfo=function(a,b){var c=a.indexOf(\"/\"),h=a.lastIndexOf(\"@\",-1<c?c:a.length-1);-1<h&&(-1===c||h<c)?(c=a.substring(0,h).split(\":\"),b.username=c[0]?d.decode(c[0]):null,\n                    c.shift(),b.password=c[0]?d.decode(c.join(\":\")):null,a=a.substring(h+1)):(b.username=null,b.password=null);return a};d.parseQuery=function(a,b){if(!a)return{};a=a.replace(/&+/g,\"&\").replace(/^\\?*&*|&+$/g,\"\");if(!a)return{};for(var c={},h=a.split(\"&\"),p=h.length,e,f,k=0;k<p;k++)if(e=h[k].split(\"=\"),f=d.decodeQuery(e.shift(),b),e=e.length?d.decodeQuery(e.join(\"=\"),b):null,r.call(c,f)){if(\"string\"===typeof c[f]||null===c[f])c[f]=[c[f]];c[f].push(e)}else c[f]=e;return c};d.build=function(a){var b=\"\";\n                a.protocol&&(b+=a.protocol+\":\");a.urn||!b&&!a.hostname||(b+=\"//\");b+=d.buildAuthority(a)||\"\";\"string\"===typeof a.path&&(\"/\"!==a.path.charAt(0)&&\"string\"===typeof a.hostname&&(b+=\"/\"),b+=a.path);\"string\"===typeof a.query&&a.query&&(b+=\"?\"+a.query);\"string\"===typeof a.fragment&&a.fragment&&(b+=\"#\"+a.fragment);return b};d.buildHost=function(a){var b=\"\";if(a.hostname)b=d.ip6_expression.test(a.hostname)?b+(\"[\"+a.hostname+\"]\"):b+a.hostname;else return\"\";a.port&&(b+=\":\"+a.port);return b};d.buildAuthority=\n                    function(a){return d.buildUserinfo(a)+d.buildHost(a)};d.buildUserinfo=function(a){var b=\"\";a.username&&(b+=d.encode(a.username),a.password&&(b+=\":\"+d.encode(a.password)),b+=\"@\");return b};d.buildQuery=function(a,b,c){var h=\"\",p,e,f,g;for(e in a)if(r.call(a,e)&&e)if(k(a[e]))for(p={},f=0,g=a[e].length;f<g;f++)void 0!==a[e][f]&&void 0===p[a[e][f]+\"\"]&&(h+=\"&\"+d.buildQueryParameter(e,a[e][f],c),!0!==b&&(p[a[e][f]+\"\"]=!0));else void 0!==a[e]&&(h+=\"&\"+d.buildQueryParameter(e,a[e],c));return h.substring(1)};\n            d.buildQueryParameter=function(a,b,c){return d.encodeQuery(a,c)+(null!==b?\"=\"+d.encodeQuery(b,c):\"\")};d.addQuery=function(a,b,c){if(\"object\"===typeof b)for(var h in b)r.call(b,h)&&d.addQuery(a,h,b[h]);else if(\"string\"===typeof b)void 0===a[b]?a[b]=c:(\"string\"===typeof a[b]&&(a[b]=[a[b]]),k(c)||(c=[c]),a[b]=(a[b]||[]).concat(c));else throw new TypeError(\"URI.addQuery() accepts an object, string as the name parameter\");};d.removeQuery=function(a,b,c){var h;if(k(b))for(c=0,h=b.length;c<h;c++)a[b[c]]=\n                    void 0;else if(\"RegExp\"===m(b))for(h in a)b.test(h)&&(a[h]=void 0);else if(\"object\"===typeof b)for(h in b)r.call(b,h)&&d.removeQuery(a,h,b[h]);else if(\"string\"===typeof b)void 0!==c?\"RegExp\"===m(c)?!k(a[b])&&c.test(a[b])?a[b]=void 0:a[b]=z(a[b],c):a[b]===c?a[b]=void 0:k(a[b])&&(a[b]=z(a[b],c)):a[b]=void 0;else throw new TypeError(\"URI.removeQuery() accepts an object, string, RegExp as the first parameter\");};d.hasQuery=function(a,b,c,h){if(\"object\"===typeof b){for(var e in b)if(r.call(b,e)&&!d.hasQuery(a,\n                            e,b[e]))return!1;return!0}if(\"string\"!==typeof b)throw new TypeError(\"URI.hasQuery() accepts an object, string as the name parameter\");switch(m(c)){case \"Undefined\":return b in a;case \"Boolean\":return a=Boolean(k(a[b])?a[b].length:a[b]),c===a;case \"Function\":return!!c(a[b],b,a);case \"Array\":return k(a[b])?(h?w:A)(a[b],c):!1;case \"RegExp\":return k(a[b])?h?w(a[b],c):!1:Boolean(a[b]&&a[b].match(c));case \"Number\":c=String(c);case \"String\":return k(a[b])?h?w(a[b],c):!1:a[b]===c;default:throw new TypeError(\"URI.hasQuery() accepts undefined, boolean, string, number, RegExp, Function as the value parameter\");\n            }};d.commonPath=function(a,b){var c=Math.min(a.length,b.length),d;for(d=0;d<c;d++)if(a.charAt(d)!==b.charAt(d)){d--;break}if(1>d)return a.charAt(0)===b.charAt(0)&&\"/\"===a.charAt(0)?\"/\":\"\";if(\"/\"!==a.charAt(d)||\"/\"!==b.charAt(d))d=a.substring(0,d).lastIndexOf(\"/\");return a.substring(0,d+1)};d.withinString=function(a,b,c){c||(c={});var h=c.start||d.findUri.start,e=c.end||d.findUri.end,f=c.trim||d.findUri.trim,k=/[a-z0-9-]=[\"']?$/i;for(h.lastIndex=0;;){var g=h.exec(a);if(!g)break;g=g.index;if(c.ignoreHtml){var l=\n                    a.slice(Math.max(g-3,0),g);if(l&&k.test(l))continue}var l=g+a.slice(g).search(e),m=a.slice(g,l).replace(f,\"\");c.ignore&&c.ignore.test(m)||(l=g+m.length,m=b(m,g,l,a),a=a.slice(0,g)+m+a.slice(l),h.lastIndex=g+m.length)}h.lastIndex=0;return a};d.ensureValidHostname=function(a){if(a.match(d.invalid_hostname_characters)){if(!f)throw new TypeError('Hostname \"'+a+'\" contains characters other than [A-Z0-9.-] and Punycode.js is not available');if(f.toASCII(a).match(d.invalid_hostname_characters))throw new TypeError('Hostname \"'+\n                    a+'\" contains characters other than [A-Z0-9.-]');}};d.noConflict=function(a){if(a)return a={URI:this.noConflict()},n.URITemplate&&\"function\"===typeof n.URITemplate.noConflict&&(a.URITemplate=n.URITemplate.noConflict()),n.IPv6&&\"function\"===typeof n.IPv6.noConflict&&(a.IPv6=n.IPv6.noConflict()),n.SecondLevelDomains&&\"function\"===typeof n.SecondLevelDomains.noConflict&&(a.SecondLevelDomains=n.SecondLevelDomains.noConflict()),a;n.URI===this&&(n.URI=D);return this};e.build=function(a){if(!0===a)this._deferred_build=\n                    !0;else if(void 0===a||this._deferred_build)this._string=d.build(this._parts),this._deferred_build=!1;return this};e.clone=function(){return new d(this)};e.valueOf=e.toString=function(){return this.build(!1)._string};e.protocol=u(\"protocol\");e.username=u(\"username\");e.password=u(\"password\");e.hostname=u(\"hostname\");e.port=u(\"port\");e.query=B(\"query\",\"?\");e.fragment=B(\"fragment\",\"#\");e.search=function(a,b){var c=this.query(a,b);return\"string\"===typeof c&&c.length?\"?\"+c:c};e.hash=function(a,b){var c=\n                    this.fragment(a,b);return\"string\"===typeof c&&c.length?\"#\"+c:c};e.pathname=function(a,b){if(void 0===a||!0===a){var c=this._parts.path||(this._parts.hostname?\"/\":\"\");return a?(this._parts.urn?d.decodeUrnPath:d.decodePath)(c):c}this._parts.path=this._parts.urn?a?d.recodeUrnPath(a):\"\":a?d.recodePath(a):\"/\";this.build(!b);return this};e.path=e.pathname;e.href=function(a,b){var c;if(void 0===a)return this.toString();this._string=\"\";this._parts=d._parts();var h=a instanceof d,e=\"object\"===typeof a&&(a.hostname||\n                    a.path||a.pathname);a.nodeName&&(e=d.getDomAttribute(a),a=a[e]||\"\",e=!1);!h&&e&&void 0!==a.pathname&&(a=a.toString());if(\"string\"===typeof a||a instanceof String)this._parts=d.parse(String(a),this._parts);else if(h||e)for(c in h=h?a._parts:a,h)r.call(this._parts,c)&&(this._parts[c]=h[c]);else throw new TypeError(\"invalid input\");this.build(!b);return this};e.is=function(a){var b=!1,c=!1,h=!1,e=!1,f=!1,g=!1,k=!1,l=!this._parts.urn;this._parts.hostname&&(l=!1,c=d.ip4_expression.test(this._parts.hostname),\n                    h=d.ip6_expression.test(this._parts.hostname),b=c||h,f=(e=!b)&&q&&q.has(this._parts.hostname),g=e&&d.idn_expression.test(this._parts.hostname),k=e&&d.punycode_expression.test(this._parts.hostname));switch(a.toLowerCase()){case \"relative\":return l;case \"absolute\":return!l;case \"domain\":case \"name\":return e;case \"sld\":return f;case \"ip\":return b;case \"ip4\":case \"ipv4\":case \"inet4\":return c;case \"ip6\":case \"ipv6\":case \"inet6\":return h;case \"idn\":return g;case \"url\":return!this._parts.urn;case \"urn\":return!!this._parts.urn;\n                case \"punycode\":return k}return null};var E=e.protocol,F=e.port,G=e.hostname;e.protocol=function(a,b){if(void 0!==a&&a&&(a=a.replace(/:(\\/\\/)?$/,\"\"),!a.match(d.protocol_expression)))throw new TypeError('Protocol \"'+a+\"\\\" contains characters other than [A-Z0-9.+-] or doesn't start with [A-Z]\");return E.call(this,a,b)};e.scheme=e.protocol;e.port=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0!==a&&(0===a&&(a=null),a&&(a+=\"\",\":\"===a.charAt(0)&&(a=a.substring(1)),a.match(/[^0-9]/))))throw new TypeError('Port \"'+\n                    a+'\" contains characters other than [0-9]');return F.call(this,a,b)};e.hostname=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0!==a){var c={};if(\"/\"!==d.parseHost(a,c))throw new TypeError('Hostname \"'+a+'\" contains characters other than [A-Z0-9.-]');a=c.hostname}return G.call(this,a,b)};e.host=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a)return this._parts.hostname?d.buildHost(this._parts):\"\";if(\"/\"!==d.parseHost(a,this._parts))throw new TypeError('Hostname \"'+\n                    a+'\" contains characters other than [A-Z0-9.-]');this.build(!b);return this};e.authority=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a)return this._parts.hostname?d.buildAuthority(this._parts):\"\";if(\"/\"!==d.parseAuthority(a,this._parts))throw new TypeError('Hostname \"'+a+'\" contains characters other than [A-Z0-9.-]');this.build(!b);return this};e.userinfo=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a){if(!this._parts.username)return\"\";var c=d.buildUserinfo(this._parts);\n                return c.substring(0,c.length-1)}\"@\"!==a[a.length-1]&&(a+=\"@\");d.parseUserinfo(a,this._parts);this.build(!b);return this};e.resource=function(a,b){var c;if(void 0===a)return this.path()+this.search()+this.hash();c=d.parse(a);this._parts.path=c.path;this._parts.query=c.query;this._parts.fragment=c.fragment;this.build(!b);return this};e.subdomain=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a){if(!this._parts.hostname||this.is(\"IP\"))return\"\";var c=this._parts.hostname.length-\n                    this.domain().length-1;return this._parts.hostname.substring(0,c)||\"\"}c=this._parts.hostname.length-this.domain().length;c=this._parts.hostname.substring(0,c);c=new RegExp(\"^\"+l(c));a&&\".\"!==a.charAt(a.length-1)&&(a+=\".\");a&&d.ensureValidHostname(a);this._parts.hostname=this._parts.hostname.replace(c,a);this.build(!b);return this};e.domain=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;\"boolean\"===typeof a&&(b=a,a=void 0);if(void 0===a){if(!this._parts.hostname||this.is(\"IP\"))return\"\";\n                var c=this._parts.hostname.match(/\\./g);if(c&&2>c.length)return this._parts.hostname;c=this._parts.hostname.length-this.tld(b).length-1;c=this._parts.hostname.lastIndexOf(\".\",c-1)+1;return this._parts.hostname.substring(c)||\"\"}if(!a)throw new TypeError(\"cannot set domain empty\");d.ensureValidHostname(a);!this._parts.hostname||this.is(\"IP\")?this._parts.hostname=a:(c=new RegExp(l(this.domain())+\"$\"),this._parts.hostname=this._parts.hostname.replace(c,a));this.build(!b);return this};e.tld=function(a,\n                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    b){if(this._parts.urn)return void 0===a?\"\":this;\"boolean\"===typeof a&&(b=a,a=void 0);if(void 0===a){if(!this._parts.hostname||this.is(\"IP\"))return\"\";var c=this._parts.hostname.lastIndexOf(\".\"),c=this._parts.hostname.substring(c+1);return!0!==b&&q&&q.list[c.toLowerCase()]?q.get(this._parts.hostname)||c:c}if(a)if(a.match(/[^a-zA-Z0-9-]/))if(q&&q.is(a))c=new RegExp(l(this.tld())+\"$\"),this._parts.hostname=this._parts.hostname.replace(c,a);else throw new TypeError('TLD \"'+a+'\" contains characters other than [A-Z0-9]');\n            else{if(!this._parts.hostname||this.is(\"IP\"))throw new ReferenceError(\"cannot set TLD on non-domain host\");c=new RegExp(l(this.tld())+\"$\");this._parts.hostname=this._parts.hostname.replace(c,a)}else throw new TypeError(\"cannot set TLD empty\");this.build(!b);return this};e.directory=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a||!0===a){if(!this._parts.path&&!this._parts.hostname)return\"\";if(\"/\"===this._parts.path)return\"/\";var c=this._parts.path.length-this.filename().length-\n                    1,c=this._parts.path.substring(0,c)||(this._parts.hostname?\"/\":\"\");return a?d.decodePath(c):c}c=this._parts.path.length-this.filename().length;c=this._parts.path.substring(0,c);c=new RegExp(\"^\"+l(c));this.is(\"relative\")||(a||(a=\"/\"),\"/\"!==a.charAt(0)&&(a=\"/\"+a));a&&\"/\"!==a.charAt(a.length-1)&&(a+=\"/\");a=d.recodePath(a);this._parts.path=this._parts.path.replace(c,a);this.build(!b);return this};e.filename=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a||!0===a){if(!this._parts.path||\n                    \"/\"===this._parts.path)return\"\";var c=this._parts.path.lastIndexOf(\"/\"),c=this._parts.path.substring(c+1);return a?d.decodePathSegment(c):c}c=!1;\"/\"===a.charAt(0)&&(a=a.substring(1));a.match(/\\.?\\//)&&(c=!0);var h=new RegExp(l(this.filename())+\"$\");a=d.recodePath(a);this._parts.path=this._parts.path.replace(h,a);c?this.normalizePath(b):this.build(!b);return this};e.suffix=function(a,b){if(this._parts.urn)return void 0===a?\"\":this;if(void 0===a||!0===a){if(!this._parts.path||\"/\"===this._parts.path)return\"\";\n                var c=this.filename(),h=c.lastIndexOf(\".\");if(-1===h)return\"\";c=c.substring(h+1);c=/^[a-z0-9%]+$/i.test(c)?c:\"\";return a?d.decodePathSegment(c):c}\".\"===a.charAt(0)&&(a=a.substring(1));if(c=this.suffix())h=a?new RegExp(l(c)+\"$\"):new RegExp(l(\".\"+c)+\"$\");else{if(!a)return this;this._parts.path+=\".\"+d.recodePath(a)}h&&(a=d.recodePath(a),this._parts.path=this._parts.path.replace(h,a));this.build(!b);return this};e.segment=function(a,b,c){var d=this._parts.urn?\":\":\"/\",e=this.path(),f=\"/\"===e.substring(0,\n                            1),e=e.split(d);void 0!==a&&\"number\"!==typeof a&&(c=b,b=a,a=void 0);if(void 0!==a&&\"number\"!==typeof a)throw Error('Bad segment \"'+a+'\", must be 0-based integer');f&&e.shift();0>a&&(a=Math.max(e.length+a,0));if(void 0===b)return void 0===a?e:e[a];if(null===a||void 0===e[a])if(k(b)){e=[];a=0;for(var g=b.length;a<g;a++)if(b[a].length||e.length&&e[e.length-1].length)e.length&&!e[e.length-1].length&&e.pop(),e.push(b[a])}else{if(b||\"string\"===typeof b)\"\"===e[e.length-1]?e[e.length-1]=b:e.push(b)}else b?\n                    e[a]=b:e.splice(a,1);f&&e.unshift(\"\");return this.path(e.join(d),c)};e.segmentCoded=function(a,b,c){var e,f;\"number\"!==typeof a&&(c=b,b=a,a=void 0);if(void 0===b){a=this.segment(a,b,c);if(k(a))for(e=0,f=a.length;e<f;e++)a[e]=d.decode(a[e]);else a=void 0!==a?d.decode(a):void 0;return a}if(k(b))for(e=0,f=b.length;e<f;e++)b[e]=d.encode(b[e]);else b=\"string\"===typeof b||b instanceof String?d.encode(b):b;return this.segment(a,b,c)};var H=e.query;e.query=function(a,b){if(!0===a)return d.parseQuery(this._parts.query,\n                    this._parts.escapeQuerySpace);if(\"function\"===typeof a){var c=d.parseQuery(this._parts.query,this._parts.escapeQuerySpace),e=a.call(this,c);this._parts.query=d.buildQuery(e||c,this._parts.duplicateQueryParameters,this._parts.escapeQuerySpace);this.build(!b);return this}return void 0!==a&&\"string\"!==typeof a?(this._parts.query=d.buildQuery(a,this._parts.duplicateQueryParameters,this._parts.escapeQuerySpace),this.build(!b),this):H.call(this,a,b)};e.setQuery=function(a,b,c){var e=d.parseQuery(this._parts.query,\n                    this._parts.escapeQuerySpace);if(\"string\"===typeof a||a instanceof String)e[a]=void 0!==b?b:null;else if(\"object\"===typeof a)for(var f in a)r.call(a,f)&&(e[f]=a[f]);else throw new TypeError(\"URI.addQuery() accepts an object, string as the name parameter\");this._parts.query=d.buildQuery(e,this._parts.duplicateQueryParameters,this._parts.escapeQuerySpace);\"string\"!==typeof a&&(c=b);this.build(!c);return this};e.addQuery=function(a,b,c){var e=d.parseQuery(this._parts.query,this._parts.escapeQuerySpace);\n                d.addQuery(e,a,void 0===b?null:b);this._parts.query=d.buildQuery(e,this._parts.duplicateQueryParameters,this._parts.escapeQuerySpace);\"string\"!==typeof a&&(c=b);this.build(!c);return this};e.removeQuery=function(a,b,c){var e=d.parseQuery(this._parts.query,this._parts.escapeQuerySpace);d.removeQuery(e,a,b);this._parts.query=d.buildQuery(e,this._parts.duplicateQueryParameters,this._parts.escapeQuerySpace);\"string\"!==typeof a&&(c=b);this.build(!c);return this};e.hasQuery=function(a,b,c){var e=d.parseQuery(this._parts.query,\n                    this._parts.escapeQuerySpace);return d.hasQuery(e,a,b,c)};e.setSearch=e.setQuery;e.addSearch=e.addQuery;e.removeSearch=e.removeQuery;e.hasSearch=e.hasQuery;e.normalize=function(){return this._parts.urn?this.normalizeProtocol(!1).normalizePath(!1).normalizeQuery(!1).normalizeFragment(!1).build():this.normalizeProtocol(!1).normalizeHostname(!1).normalizePort(!1).normalizePath(!1).normalizeQuery(!1).normalizeFragment(!1).build()};e.normalizeProtocol=function(a){\"string\"===typeof this._parts.protocol&&\n            (this._parts.protocol=this._parts.protocol.toLowerCase(),this.build(!a));return this};e.normalizeHostname=function(a){this._parts.hostname&&(this.is(\"IDN\")&&f?this._parts.hostname=f.toASCII(this._parts.hostname):this.is(\"IPv6\")&&g&&(this._parts.hostname=g.best(this._parts.hostname)),this._parts.hostname=this._parts.hostname.toLowerCase(),this.build(!a));return this};e.normalizePort=function(a){\"string\"===typeof this._parts.protocol&&this._parts.port===d.defaultPorts[this._parts.protocol]&&(this._parts.port=\n                    null,this.build(!a));return this};e.normalizePath=function(a){var b=this._parts.path;if(!b)return this;if(this._parts.urn)return this._parts.path=d.recodeUrnPath(this._parts.path),this.build(!a),this;if(\"/\"===this._parts.path)return this;var c,e=\"\",f,g;\"/\"!==b.charAt(0)&&(c=!0,b=\"/\"+b);if(\"/..\"===b.slice(-3)||\"/.\"===b.slice(-2))b+=\"/\";b=b.replace(/(\\/(\\.\\/)+)|(\\/\\.$)/g,\"/\").replace(/\\/{2,}/g,\"/\");c&&(e=b.substring(1).match(/^(\\.\\.\\/)+/)||\"\")&&(e=e[0]);for(;;){f=b.indexOf(\"/..\");if(-1===f)break;else if(0===\n                    f){b=b.substring(3);continue}g=b.substring(0,f).lastIndexOf(\"/\");-1===g&&(g=f);b=b.substring(0,g)+b.substring(f+3)}c&&this.is(\"relative\")&&(b=e+b.substring(1));b=d.recodePath(b);this._parts.path=b;this.build(!a);return this};e.normalizePathname=e.normalizePath;e.normalizeQuery=function(a){\"string\"===typeof this._parts.query&&(this._parts.query.length?this.query(d.parseQuery(this._parts.query,this._parts.escapeQuerySpace)):this._parts.query=null,this.build(!a));return this};e.normalizeFragment=function(a){this._parts.fragment||\n            (this._parts.fragment=null,this.build(!a));return this};e.normalizeSearch=e.normalizeQuery;e.normalizeHash=e.normalizeFragment;e.iso8859=function(){var a=d.encode,b=d.decode;d.encode=escape;d.decode=decodeURIComponent;try{this.normalize()}finally{d.encode=a,d.decode=b}return this};e.unicode=function(){var a=d.encode,b=d.decode;d.encode=x;d.decode=unescape;try{this.normalize()}finally{d.encode=a,d.decode=b}return this};e.readable=function(){var a=this.clone();a.username(\"\").password(\"\").normalize();\n                var b=\"\";a._parts.protocol&&(b+=a._parts.protocol+\"://\");a._parts.hostname&&(a.is(\"punycode\")&&f?(b+=f.toUnicode(a._parts.hostname),a._parts.port&&(b+=\":\"+a._parts.port)):b+=a.host());a._parts.hostname&&a._parts.path&&\"/\"!==a._parts.path.charAt(0)&&(b+=\"/\");b+=a.path(!0);if(a._parts.query){for(var c=\"\",e=0,g=a._parts.query.split(\"&\"),k=g.length;e<k;e++){var l=(g[e]||\"\").split(\"=\"),c=c+(\"&\"+d.decodeQuery(l[0],this._parts.escapeQuerySpace).replace(/&/g,\"%26\"));void 0!==l[1]&&(c+=\"=\"+d.decodeQuery(l[1],\n                                this._parts.escapeQuerySpace).replace(/&/g,\"%26\"))}b+=\"?\"+c.substring(1)}return b+=d.decodeQuery(a.hash(),!0)};e.absoluteTo=function(a){var b=this.clone(),c=[\"protocol\",\"username\",\"password\",\"hostname\",\"port\"],e,f;if(this._parts.urn)throw Error(\"URNs do not have any generally defined hierarchical components\");a instanceof d||(a=new d(a));b._parts.protocol||(b._parts.protocol=a._parts.protocol);if(this._parts.hostname)return b;for(e=0;f=c[e];e++)b._parts[f]=a._parts[f];b._parts.path?\"..\"===b._parts.path.substring(-2)&&\n            (b._parts.path+=\"/\"):(b._parts.path=a._parts.path,b._parts.query||(b._parts.query=a._parts.query));\"/\"!==b.path().charAt(0)&&(c=(c=a.directory())?c:0===a.path().indexOf(\"/\")?\"/\":\"\",b._parts.path=(c?c+\"/\":\"\")+b._parts.path,b.normalizePath());b.build();return b};e.relativeTo=function(a){var b=this.clone().normalize(),c,e,f;if(b._parts.urn)throw Error(\"URNs do not have any generally defined hierarchical components\");a=(new d(a)).normalize();c=b._parts;e=a._parts;f=b.path();a=a.path();if(\"/\"!==f.charAt(0))throw Error(\"URI is already relative\");\n                if(\"/\"!==a.charAt(0))throw Error(\"Cannot calculate a URI relative to another relative URI\");c.protocol===e.protocol&&(c.protocol=null);if(c.username===e.username&&c.password===e.password&&null===c.protocol&&null===c.username&&null===c.password&&c.hostname===e.hostname&&c.port===e.port)c.hostname=null,c.port=null;else return b.build();if(f===a)return c.path=\"\",b.build();f=d.commonPath(f,a);if(!f)return b.build();e=e.path.substring(f.length).replace(/[^\\/]*$/,\"\").replace(/.*?\\//g,\"../\");c.path=e+c.path.substring(f.length)||\n                        \"./\";return b.build()};e.equals=function(a){var b=this.clone();a=new d(a);var c={},e={},f={},g;b.normalize();a.normalize();if(b.toString()===a.toString())return!0;c=b.query();e=a.query();b.query(\"\");a.query(\"\");if(b.toString()!==a.toString()||c.length!==e.length)return!1;c=d.parseQuery(c,this._parts.escapeQuerySpace);e=d.parseQuery(e,this._parts.escapeQuerySpace);for(g in c)if(r.call(c,g)){if(!k(c[g])){if(c[g]!==e[g])return!1}else if(!A(c[g],e[g]))return!1;f[g]=!0}for(g in e)if(r.call(e,g)&&!f[g])return!1;\n                return!0};e.duplicateQueryParameters=function(a){this._parts.duplicateQueryParameters=!!a;return this};e.escapeQuerySpace=function(a){this._parts.escapeQuerySpace=!!a;return this};return d});\n        (function(f,g){\"object\"===typeof exports?module.exports=g(require(\"./URI\")):\"function\"===typeof define&&define.amd?define([\"./URI\"],g):g(f.URI)})(this,function(f){var g=f.prototype,q=g.fragment;f.fragmentPrefix=\"?\";var n=f._parts;f._parts=function(){var d=n();d.fragmentPrefix=f.fragmentPrefix;return d};g.fragmentPrefix=function(d){this._parts.fragmentPrefix=d;return this};g.fragment=function(d,g){var m=this._parts.fragmentPrefix,k=this._parts.fragment||\"\";return!0===d?k.substring(0,m.length)!==m?\n        {}:f.parseQuery(k.substring(m.length)):void 0!==d&&\"string\"!==typeof d?(this._parts.fragment=m+f.buildQuery(d),this.build(!g),this):q.call(this,d,g)};g.addFragment=function(d,g,m){var k=this._parts.fragmentPrefix,n=f.parseQuery((this._parts.fragment||\"\").substring(k.length));f.addQuery(n,d,g);this._parts.fragment=k+f.buildQuery(n);\"string\"!==typeof d&&(m=g);this.build(!m);return this};g.removeFragment=function(d,g,m){var k=this._parts.fragmentPrefix,n=f.parseQuery((this._parts.fragment||\"\").substring(k.length));\n            f.removeQuery(n,d,g);this._parts.fragment=k+f.buildQuery(n);\"string\"!==typeof d&&(m=g);this.build(!m);return this};g.addHash=g.addFragment;g.removeHash=g.removeFragment;return f});\n    </script>\n";
 
 const tail = `
-{{scripts}}
+{{ scripts }}
 
 </head>
 <body>
